requirements.txt
# Core dependencies
python-dotenv==1.0.0
requests==2.31.0

# Reddit API
praw==7.7.1

# Data processing
pandas==2.1.4
numpy==1.26.2

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Development tools
pytest==7.4.3
black==23.12.1
flake8==7.0.0